refactor: response 패키지의 오타 수정 및 관련 임포트 모두 수정
