+++ src/main/java/org/example/pedia_777/domain/movie/repository/MovieRepositoryCustom.java
@@ -1,6 +1,6 @@
 package org.example.pedia_777.domain.movie.repository;
 
-import org.example.pedia_777.domain.search.dto.respoonse.MovieSearchProjection;
+import org.example.pedia_777.domain.search.dto.response.MovieSearchProjection;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
