+++ src/main/java/org/example/pedia_777/domain/movie/service/MovieServiceApi.java
@@ -1,7 +1,7 @@
 package org.example.pedia_777.domain.movie.service;
 
 import org.example.pedia_777.domain.movie.entity.Movie;
-import org.example.pedia_777.domain.search.dto.respoonse.MovieSearchResponse;
+import org.example.pedia_777.domain.search.dto.response.MovieSearchResponse;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
