+++ src/main/java/org/example/pedia_777/domain/search/dto/response/MovieSearchProjection.java
@@ -1,4 +1,4 @@
-package org.example.pedia_777.domain.search.dto.respoonse;
+package org.example.pedia_777.domain.search.dto.response;
 
 public record MovieSearchProjection(
         Long id,
