+++ src/main/java/org/example/pedia_777/domain/search/service/SearchService.java
@@ -3,7 +3,7 @@ package org.example.pedia_777.domain.search.service;
 import lombok.RequiredArgsConstructor;
 import org.example.pedia_777.common.dto.PageResponse;
 import org.example.pedia_777.domain.movie.service.MovieService;
-import org.example.pedia_777.domain.search.dto.respoonse.MovieSearchResponse;
+import org.example.pedia_777.domain.search.dto.response.MovieSearchResponse;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
